% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin_exec.R
\name{reinstall_and_test_addin}
\alias{reinstall_and_test_addin}
\title{Reinstall and Test the pcir Package}
\usage{
reinstall_and_test_addin()
}
\description{
This RStudio Addin reinstalls the 'pcir' package from the local source,
reloads it, and runs all functions on example data for testing.
}
