% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubble.R, R/pcir.package.R
\name{bubble}
\alias{bubble}
\title{Create a Bubble Plot for PCI Visualization}
\usage{
bubble(df3)

bubble(df3)
}
\arguments{
\item{df3}{A data frame generated by the \code{pci} function.}
}
\value{
A ggplot2 object representing the bubble plot.

A ggplot2 object representing the bubble plot.
}
\description{
This function generates a bubble plot to visualize the results of the PCI
calculation. It shows the mean action acceptability on the y-axis and the PCI
value as the size of the bubbles.

Create a bubble plot to visualize PCI results.
}
\examples{
data <- data.frame(
  Category = rep(c('A', 'B', 'C'), each = 3),
  Value = c(-1, 0, 1, -1, 0, 1, -1, 0, 1)
)
df2 <- counting(data)
df3 <- pci(df2)
bubble(df3)
df3 <- pci(df2)
bubble(df3)
}
